+++ packages/odyssey-storybook/src/components/odyssey-mui/Typography/Typography.stories.tsx
@@ -89,6 +89,10 @@ const storybookMeta: Meta<TypographyProps> = {
           summary: "ReactNode",
         },
       },
+      type: {
+        required: true,
+        name: "string",
+      },
     },
     color: {
       options: typographyColorValues,